Better bash prompt and bash file snippet
Improved the bash prompt by manually adding info on conda virtual
environment and kubectl context. Dramatically improved the skeleton
snippet for bash script files. Fixed a setting in indentLine plugin that
forced JSON and Markdown files to hide syntax elements. Deleted a broken
plugin from tmux that was supposed to show when prefix key was active.
Started tracking a select subset of executables from the home folder
(for example, "brightness" is a new command to set screen brightness,
and "tmux-project-template" is a template file for pre-arranging tmux
windows for a new session).