Python code snippets
Added a few python code snippets, similar to the UltiSnips setup in
Vim.  Adjusted pyright settings to improve autocompletions.

Added a convenience command for launching and stopping Kubernetes.
Disabled Kubernetes and Docker in systemd by default to save CPU and
battery power.

Also fixed a bug that prevented interactive code execution in ipython
and a bug that sometimes caused errors in treemacs when switching tabs.