Improved smartparens and vterm functionality
Improved smartparens settings and behaviour, including situations when
newlines are part of the pair and when tick boxes are added in org mode.

Refactored the paren formatting function from the python module to make
it more efficient and more generic. It can now be used in JS as well.

Improved bindings for vterm which now supports pasting from clipboard
and behaves correctly when pasting text in normal state.