Switched to straight.el for package management
Switched to straight.el for package management.

Rebuilt emacs using latest version from github because previous version
caused an error with straight.el at initialization.

Deprecated EAF widgets as they seem to freeze when trying to pull with
straight.el and I don't seem to be using them anyway.

Fixed a bug that caused windows to disappear when a vterm buffer was
being created.

Fixed a bug in copy-paste keybindings in vterm-mode that caused them to
spill into other modes.