Stable mu4e release + a few useful Emacs config overrides
Switched to the stable mu4e branch 1.6 which seems less buggy. For
example, the setting for jumping directly to the contents of an email
message via an org link ()rather than the header view) seems to work in
the 1.6 branch but not in the dev branch.

Also added a few useful settings overrides, such as saving and
retrieving minibuffer history - curtesy of:

  [https://www.youtube.com/watch?v=51eSeqcaikM]