Switching from OpenVPN to Wireguard; minor ergonomic changes
Switched from OpenVPN to Wireguard as the main VPN server technology.
Openvpn integration in Emacs has been disabled. Bash aliases have been
updated to spin up wireguard. Also added bash aliases for checking
public IP from command line.

Improved the command for shutting down local Kubernetes cluster. New
approach drains the node instead of attempting to shut down the daemons
which results in a more graceful approach. However, not all pods seem to
be terminating so some further modifications may still be required.

Deleted the Emacs code snippet which mounted the project drive on Emacs
startup. All drives are currently encrypted and the added security
benefit of not auto-mounting the project drive on startup is minimal.
Furthermore, the container storage location for containerd had to be
moved to the project drive to conserve space for the local kubelet
process, which means mounting the project drive on startup is now the
more graceful approach to running Kubernetes.

Finally added a shortcut for applying/deleting YAML manifest files in
Kubernetes directly from the manifest buffer. Previous approach was to
spin up a shell in the file's directory and apply using the command line
which was messy.

Added a placeholder for an alternative approach to IDE functionality in
Emacs (vertigo / corfu / marginalia). Current setup with ivy / counsel /
company seems to work well so no urgency in applying the change, but
something to consider for the future.

Finally fixed bug which made LSP UI doc frames appear with a tab-bar.

Disabled backup files in an attempt to further / better rely on
undo-tree. Unfortunately, still getting occasional undo-tree file errors
and trying to figure out where those come from.